# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UpFlow - Regulação de Leitos API",
        "description": "API REST para gerenciamento e controle de leitos hospitalares desenvolvida em Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Ambiente Local"
        }
    ],
    "paths": {
...
        "/api/pacientes": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Lista todos os pacientes",
                "operationId": "9cf5506c2b465d4386c71237b6d290d1",
                "responses": {
                    "200": {
                        "description": "Lista de pacientes retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Cadastra um novo paciente",
                "operationId": "ffd18f87de3d0cd123142caba8dfbaf1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "data_nascimento",
                                    "cpf"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Ana Oliveira"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-03-20"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "22233344455"
                                    },
                                    "cartao_sus": {
                                        "type": "string",
                                        "example": "222333444555666"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paciente cadastrado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/pacientes/{paciente}": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Exibe um paciente específico",
                "operationId": "b83e407954cb7f98e027358bd8711df8",
                "parameters": [
                    {
                        "name": "paciente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente retornado com sucesso"
                    },
                    "404": {
                        "description": "Paciente não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Atualiza um paciente",
                "operationId": "f287c4047ef2f1aa8a8299d1185dd343",
                "parameters": [
                    {
                        "name": "paciente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Ana Oliveira"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-03-20"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "22233344455"
                                    },
                                    "cartao_sus": {
                                        "type": "string",
                                        "example": "222333444555666"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Remove um paciente",
                "operationId": "d5958db346d596bd4c8993d220b8c38c",
                "parameters": [
                    {
                        "name": "paciente",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Paciente removido com sucesso"
                    },
                    "404": {
                        "description": "Paciente não encontrado"
                    }
                }
            }
        },
        "/api/pacientes/cpf/{cpf}/leito": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Busca o leito atual de um paciente pelo CPF",
                "description": "Retorna o paciente, a internação ativa, o leito ocupado e o tipo de leito.",
                "operationId": "f9ab71b0f588c0346108cc247ffb533c",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "11122233344"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leito do paciente retornado com sucesso"
                    },
                    "422": {
                        "description": "Paciente não encontrado ou sem internação ativa"
                    }
                }
            }
        },
...
    },
...
}